refactor: using axios cjs
