+++ packages/daemon/src/types/axios.d.ts
@@ -1,0 +1,4 @@
+/**
+ * Declaration file for axios CommonJS module
+ */
+declare module 'axios/dist/node/axios.cjs';
